feat(plugins): replace ayamir/project.nvim with DrKJeff16/project.nvim
ive been using this fork for a while now, and since the author acc put in the
work to port over a bunch of the stale PRs, it just makes more sense to use it.
Plus this way @ayamir doesn't have to keep maintaining it just to keep stuff
working too!

Signed-off-by: Jint-lzxy <50296129+Jint-lzxy@users.noreply.github.com>